# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Accounting app API specification
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface Account {
  id: number;
  name: string;
  bankName: string;
  accountNumber: string;
  emoji?: string;
  /** @nullable */
  imageUrl?: string | null;
  balance: number;
  createdAt: string;
}

export interface AccountInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  bankName: string;
  /** @minLength 1 */
  accountNumber: string;
  emoji?: string;
  /** @nullable */
  imageUrl?: string | null;
  initialBalance?: number;
}

export interface AccountUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  bankName?: string;
  /** @minLength 1 */
  accountNumber?: string;
  emoji?: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface Subcategory {
  id: number;
  categoryId: number;
  name: string;
  emoji: string;
}

export interface Category {
  id: number;
  name: string;
  emoji: string;
  /** @nullable */
  budget?: number | null;
  /**
     * budget minus expenses plus deposits this month
     * @nullable
     */
  currentBalance?: number | null;
  /** @nullable */
  accountId?: number | null;
  /** @nullable */
  accountName?: string | null;
  subcategories: Subcategory[];
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  emoji: string;
  /** @nullable */
  budget?: number | null;
  /** @nullable */
  accountId?: number | null;
}

export interface CategoryUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  emoji?: string;
  /** @nullable */
  budget?: number | null;
  /** @nullable */
  accountId?: number | null;
}

export interface SubcategoryInput {
  categoryId: number;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  emoji: string;
}

export interface SubcategoryUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  emoji?: string;
}

export interface Salary {
  id: number;
  amount: number;
  depositDay: number;
  /** @nullable */
  accountId?: number | null;
  /** @nullable */
  accountName?: string | null;
  /** @nullable */
  notes?: string | null;
  updatedAt: string;
}

export interface SalaryInput {
  /** @minimum 0 */
  amount: number;
  /**
     * @minimum 1
     * @maximum 31
     */
  depositDay: number;
  /** @nullable */
  accountId?: number | null;
  notes?: string;
}

export interface SalaryAllocation {
  id: number;
  categoryId: number;
  categoryName?: string;
  categoryEmoji?: string;
  amount: number;
}

export interface SalaryAllocationInput {
  categoryId: number;
  /** @minimum 0 */
  amount: number;
}

export interface SalaryAllocationUpdate {
  /** @minimum 0 */
  amount?: number;
}

export interface SalaryProcessResult {
  processed: boolean;
  alreadyProcessed?: boolean;
  message: string;
}

export interface Loan {
  id: number;
  name: string;
  totalAmount: number;
  monthlyInstallment: number;
  months: number;
  startDate: string;
  remainingMonths: number;
  isActive: boolean;
  createdAt: string;
}

export interface LoanInput {
  /** @minLength 1 */
  name: string;
  /** @minimum 0 */
  totalAmount: number;
  /** @minimum 0 */
  monthlyInstallment: number;
  /** @minimum 1 */
  months: number;
  startDate: string;
  /** @minimum 0 */
  remainingMonths?: number;
}

export interface LoanUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minimum 0 */
  totalAmount?: number;
  /** @minimum 0 */
  monthlyInstallment?: number;
  /** @minimum 1 */
  months?: number;
  /** @minimum 0 */
  remainingMonths?: number;
  isActive?: boolean;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  expense: 'expense',
  deposit: 'deposit',
} as const;

export interface Transaction {
  id: number;
  type: TransactionType;
  amount: number;
  date: string;
  accountId: number;
  /** @nullable */
  subcategoryId?: number | null;
  /** @nullable */
  subcategoryName?: string | null;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export type TransactionInputType = typeof TransactionInputType[keyof typeof TransactionInputType];


export const TransactionInputType = {
  expense: 'expense',
  deposit: 'deposit',
} as const;

export interface TransactionInput {
  type: TransactionInputType;
  /** @minimum 0.01 */
  amount: number;
  date: string;
  accountId: number;
  /** @nullable */
  subcategoryId?: number | null;
  /** @nullable */
  notes?: string | null;
}

export type TransactionUpdateType = typeof TransactionUpdateType[keyof typeof TransactionUpdateType];


export const TransactionUpdateType = {
  expense: 'expense',
  deposit: 'deposit',
} as const;

export interface TransactionUpdate {
  type?: TransactionUpdateType;
  /** @minimum 0.01 */
  amount?: number;
  date?: string;
  accountId?: number;
  /** @nullable */
  subcategoryId?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface AccountBalance {
  accountId: number;
  name: string;
  emoji?: string;
  /** @nullable */
  imageUrl?: string | null;
  bankName: string;
  balance: number;
}

export interface DashboardSummary {
  totalBalance: number;
  monthlyExpenses: number;
  monthlyDeposits: number;
  /** @nullable */
  salary: number | null;
  /** @nullable */
  depositDay?: number | null;
  currentMonth: string;
  accountBalances: AccountBalance[];
}

export interface SubcategorySpending {
  subcategoryId: number;
  name: string;
  emoji: string;
  total: number;
}

export interface CategorySpending {
  categoryId: number;
  categoryName: string;
  emoji: string;
  total: number;
  subcategories: SubcategorySpending[];
}

export type StatementEntryType = typeof StatementEntryType[keyof typeof StatementEntryType];


export const StatementEntryType = {
  expense: 'expense',
  deposit: 'deposit',
} as const;

export interface StatementEntry {
  id: number;
  type: StatementEntryType;
  amount: number;
  date: string;
  /** @nullable */
  subcategoryName?: string | null;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  notes?: string | null;
  runningBalance: number;
}

export interface AccountStatement {
  account: Account;
  openingBalance: number;
  closingBalance: number;
  transactions: StatementEntry[];
}

export type ListCategoriesParams = {
/**
 * Month for balance calculation (YYYY-MM), defaults to current month
 * @nullable
 */
month?: string | null;
};

export type ListTransactionsParams = {
/**
 * @nullable
 */
accountId?: number | null;
/**
 * @nullable
 */
type?: ListTransactionsType;
/**
 * Filter by month (YYYY-MM)
 * @nullable
 */
month?: string | null;
};

export type ListTransactionsType = typeof ListTransactionsType[keyof typeof ListTransactionsType] | null;


export const ListTransactionsType = {
  expense: 'expense',
  deposit: 'deposit',
} as const;

export type GetSpendingByCategoryParams = {
/**
 * Month to summarize (YYYY-MM), defaults to current month
 * @nullable
 */
month?: string | null;
};

export type GetAccountStatementParams = {
accountId: number;
/**
 * @nullable
 */
month?: string | null;
};