# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Accounting app API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountInput"
      responses:
        "201":
          description: Created account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get a single account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountUpdate"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all main categories with sub-categories and balances
      parameters:
        - name: month
          in: query
          description: Month for balance calculation (YYYY-MM), defaults to current month
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a main category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a main category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a main category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /subcategories:
    post:
      operationId: createSubcategory
      tags: [categories]
      summary: Create a sub-category under a main category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubcategoryInput"
      responses:
        "201":
          description: Created sub-category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subcategory"
  /subcategories/{id}:
    patch:
      operationId: updateSubcategory
      tags: [categories]
      summary: Update a sub-category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubcategoryUpdate"
      responses:
        "200":
          description: Updated sub-category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subcategory"
    delete:
      operationId: deleteSubcategory
      tags: [categories]
      summary: Delete a sub-category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /loans:
    get:
      operationId: listLoans
      tags: [loans]
      summary: List all loans
      responses:
        "200":
          description: List of loans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Loan"
    post:
      operationId: createLoan
      tags: [loans]
      summary: Create a new loan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanInput"
      responses:
        "201":
          description: Created loan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{id}:
    patch:
      operationId: updateLoan
      tags: [loans]
      summary: Update a loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanUpdate"
      responses:
        "200":
          description: Updated loan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
    delete:
      operationId: deleteLoan
      tags: [loans]
      summary: Delete a loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Account:
      type: object
      required: [id, name, bankName, accountNumber, balance, createdAt]
      properties:
        id:            { type: integer }
        name:          { type: string }
        bankName:      { type: string }
        accountNumber: { type: string }
        emoji:         { type: string }
        imageUrl:      { type: ["string", "null"] }
        balance:       { type: number }
        createdAt:     { type: string, format: date-time }
    AccountInput:
      type: object
      required: [name, bankName, accountNumber]
      properties:
        name:           { type: string, minLength: 1 }
        bankName:       { type: string, minLength: 1 }
        accountNumber:  { type: string, minLength: 1 }
        emoji:          { type: string }
        imageUrl:       { type: ["string", "null"] }
        initialBalance: { type: number, default: 0 }
    AccountUpdate:
      type: object
      properties:
        name:          { type: string, minLength: 1 }
        bankName:      { type: string, minLength: 1 }
        accountNumber: { type: string, minLength: 1 }
        emoji:         { type: string }
        imageUrl:      { type: ["string", "null"] }
    Category:
      type: object
      required: [id, name, emoji, subcategories]
      properties:
        id:             { type: integer }
        name:           { type: string }
        emoji:          { type: string }
        budget:         { type: ["number", "null"] }
        currentBalance: { type: ["number", "null"], description: "budget minus expenses plus deposits this month" }
        accountId:      { type: ["integer", "null"] }
        accountName:    { type: ["string", "null"] }
        subcategories:
          type: array
          items:
            $ref: "#/components/schemas/Subcategory"
    CategoryInput:
      type: object
      required: [name, emoji]
      properties:
        name:      { type: string, minLength: 1 }
        emoji:     { type: string, minLength: 1 }
        budget:    { type: ["number", "null"] }
        accountId: { type: ["integer", "null"] }
    CategoryUpdate:
      type: object
      properties:
        name:      { type: string, minLength: 1 }
        emoji:     { type: string, minLength: 1 }
        budget:    { type: ["number", "null"] }
        accountId: { type: ["integer", "null"] }
    Subcategory:
      type: object
      required: [id, categoryId, name, emoji]
      properties:
        id:         { type: integer }
        categoryId: { type: integer }
        name:       { type: string }
        emoji:      { type: string }
    SubcategoryInput:
      type: object
      required: [categoryId, name, emoji]
      properties:
        categoryId: { type: integer }
        name:       { type: string, minLength: 1 }
        emoji:      { type: string, minLength: 1 }
    SubcategoryUpdate:
      type: object
      properties:
        name:  { type: string, minLength: 1 }
        emoji: { type: string, minLength: 1 }
...
    Loan:
      type: object
      required: [id, name, totalAmount, monthlyInstallment, months, startDate, remainingMonths, isActive, createdAt]
      properties:
        id:                 { type: integer }
        name:               { type: string }
        totalAmount:        { type: number }
        monthlyInstallment: { type: number }
        months:             { type: integer }
        startDate:          { type: string, format: date }
        remainingMonths:    { type: integer }
        isActive:           { type: boolean }
        createdAt:          { type: string, format: date-time }
    LoanInput:
      type: object
      required: [name, totalAmount, monthlyInstallment, months, startDate]
      properties:
        name:               { type: string, minLength: 1 }
        totalAmount:        { type: number, minimum: 0 }
        monthlyInstallment: { type: number, minimum: 0 }
        months:             { type: integer, minimum: 1 }
        startDate:          { type: string, format: date }
        remainingMonths:    { type: integer, minimum: 0 }
    LoanUpdate:
      type: object
      properties:
        name:               { type: string, minLength: 1 }
        totalAmount:        { type: number, minimum: 0 }
        monthlyInstallment: { type: number, minimum: 0 }
        months:             { type: integer, minimum: 1 }
        remainingMonths:    { type: integer, minimum: 0 }
        isActive:           { type: boolean }
...