# lib/api-zod/src/generated/types/accountInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Accounting app API specification
 * OpenAPI spec version: 0.2.0
 */

export interface AccountInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  bankName: string;
  /** @minLength 1 */
  accountNumber: string;
  emoji?: string;
  /** @nullable */
  imageUrl?: string | null;
  initialBalance?: number;
}