# lib/api-zod/src/generated/types/categoryUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Accounting app API specification
 * OpenAPI spec version: 0.2.0
 */

export interface CategoryUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  emoji?: string;
  /** @nullable */
  budget?: number | null;
  /** @nullable */
  accountId?: number | null;
}