# lib/api-zod/src/generated/types/loanUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Accounting app API specification
 * OpenAPI spec version: 0.2.0
 */

export interface LoanUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minimum 0 */
  totalAmount?: number;
  /** @minimum 0 */
  monthlyInstallment?: number;
  /** @minimum 1 */
  months?: number;
  /** @minimum 0 */
  remainingMonths?: number;
  isActive?: boolean;
}