# lib/api-zod/src/generated/types/subcategory.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Accounting app API specification
 * OpenAPI spec version: 0.2.0
 */

export interface Subcategory {
  id: number;
  categoryId: number;
  name: string;
  emoji: string;
}